pyproject.toml
[project]
name = "app"
version = "1.0.0"
description = ""
authors = []
readme = "README.md"
requires-python = ">=3.10,<3.15"
dependencies = [
    "django-environ (>=0.12.0,<0.13.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "django (>=5.2.6,<6.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "django-allauth (>=65.11.2,<66.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "oauthlib (>=3.3.1,<4.0.0)",
    "django-cleanup (>=9.0.0,<10.0.0)",
    "b2sdk (>=2.10.0,<3.0.0)",
    "celery[redis] (>=5.5.3,<6.0.0)",
    "django-auditlog (>=3.2.1,<4.0.0)",
    "django-log-viewer (>=1.1.8,<2.0.0)",
    "django-unfold (>=0.67.0,<0.68.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)",
    "geoip2 (>=5.1.0,<6.0.0)",
    "torch (>=2.9.0,<3.0.0)",
    "transformers (>=4.57.1,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
faker = "^37.6.0"

